agents/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
web3>=6.15.0
python-dotenv>=1.0.0
eth-abi>=4.2.0

